Changed oids to be psuedo random positive longs (63 bits). This gets rid of a synchronisation bottleneck in the store. Consider making oids to be bucket specific in future.
Former-commit-id: 81130cd44ca61885230c924f421e6fc6e2c0318e

+++ jstore/src/main/java/uk/ac/standrews/cs/digitising_scotland/jstore/impl/BucketBackedInputStream.java
@@ -43,7 +43,7 @@ public class BucketBackedInputStream<T extends ILXP> implements IInputStream<T> 
         public T next() {
 
             try {
-                return bucket.get(Integer.parseInt(file_iterator.next().getName()));
+                return bucket.get(Long.parseLong(file_iterator.next().getName()));
 
             } catch (BucketException e) {
                 ErrorHandling.exceptionError(e, "Exception in iterator");
