+++ jstore/src/main/java/uk/ac/standrews/cs/digitising_scotland/jstore/interfaces/ILXPFactory.java
@@ -15,15 +15,15 @@ public interface ILXPFactory<T extends ILXP> {
     /*
      * create an instance of a <T extends ILXP> from the reader
      */
-    T create(int persistent_object_id, JSONReader reader) throws PersistentObjectException;
+    T create(long persistent_object_id, JSONReader reader) throws PersistentObjectException;
 
     /*
      * return true if the factory can create an instance with the specified label and false otherwise
      */
-    boolean checkConsistentWith(int label_id) throws IOException, PersistentObjectException;
+    boolean checkConsistentWith(long label_id) throws IOException, PersistentObjectException;
 
     /*
      * return the label id required by type T
      */
-    int getTypeLabel();
+    long getTypeLabel();
 }
