+++ linkage/src/main/java/uk/ac/standrews/cs/digitising_scotland/linkage/lxp_records/Marriage.java
@@ -99,7 +99,7 @@ public class Marriage extends AbstractLXP {
         super();
     }
 
-    public Marriage(int label_id, JSONReader reader) throws PersistentObjectException {
+    public Marriage(long label_id, JSONReader reader) throws PersistentObjectException {
 
         super(Store.getInstance().getNextFreePID(), reader);
     }
